Moved/renamed endpointer.go to common/client.go.
Renamed EndpointerClient to Client.
