+++ clients/common/client.go
@@ -12,7 +12,7 @@
  * the License.
  *******************************************************************************/
 
-package clients
+package common
 
 import (
 	"errors"
@@ -22,7 +22,7 @@ import (
 	"github.com/edgexfoundry/go-mod-core-contracts/clients/types"
 )
 
-type EndpointerClient struct {
+type Client struct {
 	url         string
 	initialized bool
 	endpoint    interfaces.Endpointer
@@ -30,10 +30,10 @@ type EndpointerClient struct {
 
 var notYetInitialized = errors.New("client not yet initialized")
 
-// NewEndpointerClient returns a pointer to a EndpointerClient.
+// NewClient returns a pointer to a Client.
 // A pointer is used so that when using configuration from a registry, the URL can be updated asynchronously.
-func NewEndpointerClient(params types.EndpointParams, m interfaces.Endpointer) *EndpointerClient {
-	d := EndpointerClient{initialized: false, endpoint: m}
+func NewClient(params types.EndpointParams, m interfaces.Endpointer) *Client {
+	d := Client{initialized: false, endpoint: m}
 	d.init(params)
 
 	return &d
@@ -41,7 +41,7 @@ func NewEndpointerClient(params types.EndpointParams, m interfaces.Endpointer) *
 
 // URL calls URL for timeout seconds. If a value is loaded in that time, it returns it.
 // Otherwise, it returns an error.
-func (e *EndpointerClient) URL(timeout int) (string, error) {
+func (e *Client) URL(timeout int) (string, error) {
 	timer := time.After(time.Duration(timeout) * time.Second)
 	ticker := time.Tick(500 * time.Millisecond)
 	for {
@@ -57,7 +57,7 @@ func (e *EndpointerClient) URL(timeout int) (string, error) {
 	}
 }
 
-func (e *EndpointerClient) init(params types.EndpointParams) {
+func (e *Client) init(params types.EndpointParams) {
 	if params.UseRegistry {
 		go func(ch chan string) {
 			for {
