+++ clients/metadata/device.go
@@ -21,6 +21,7 @@ import (
 	"strconv"
 
 	"github.com/edgexfoundry/go-mod-core-contracts/clients"
+	"github.com/edgexfoundry/go-mod-core-contracts/clients/common"
 	"github.com/edgexfoundry/go-mod-core-contracts/clients/interfaces"
 	"github.com/edgexfoundry/go-mod-core-contracts/clients/types"
 	"github.com/edgexfoundry/go-mod-core-contracts/models"
@@ -75,12 +76,12 @@ type DeviceClient interface {
 }
 
 type deviceRestClient struct {
-	endpointer *clients.EndpointerClient
+	client *common.Client
 }
 
 // NewDeviceClient creates an instance of DeviceClient
 func NewDeviceClient(params types.EndpointParams, m interfaces.Endpointer) DeviceClient {
-	return &deviceRestClient{endpointer: clients.NewEndpointerClient(params, m)}
+	return &deviceRestClient{client: common.NewClient(params, m)}
 }
 
 // Helper method to request and decode a device
@@ -108,7 +109,7 @@ func (d *deviceRestClient) requestDeviceSlice(url string, ctx context.Context) (
 }
 
 func (d *deviceRestClient) CheckForDevice(token string, ctx context.Context) (models.Device, error) {
-	deviceURL, err := d.endpointer.URL(10)
+	deviceURL, err := d.client.URL(10)
 	if err != nil {
 		return models.Device{}, err
 	}
@@ -117,7 +118,7 @@ func (d *deviceRestClient) CheckForDevice(token string, ctx context.Context) (mo
 }
 
 func (d *deviceRestClient) Device(id string, ctx context.Context) (models.Device, error) {
-	deviceURL, err := d.endpointer.URL(10)
+	deviceURL, err := d.client.URL(10)
 	if err != nil {
 		return models.Device{}, err
 	}
@@ -126,7 +127,7 @@ func (d *deviceRestClient) Device(id string, ctx context.Context) (models.Device
 }
 
 func (d *deviceRestClient) Devices(ctx context.Context) ([]models.Device, error) {
-	deviceURL, err := d.endpointer.URL(10)
+	deviceURL, err := d.client.URL(10)
 	if err != nil {
 		return nil, err
 	}
@@ -135,7 +136,7 @@ func (d *deviceRestClient) Devices(ctx context.Context) ([]models.Device, error)
 }
 
 func (d *deviceRestClient) DeviceForName(name string, ctx context.Context) (models.Device, error) {
-	deviceURL, err := d.endpointer.URL(10)
+	deviceURL, err := d.client.URL(10)
 	if err != nil {
 		return models.Device{}, err
 	}
@@ -144,7 +145,7 @@ func (d *deviceRestClient) DeviceForName(name string, ctx context.Context) (mode
 }
 
 func (d *deviceRestClient) DevicesByLabel(label string, ctx context.Context) ([]models.Device, error) {
-	deviceURL, err := d.endpointer.URL(10)
+	deviceURL, err := d.client.URL(10)
 	if err != nil {
 		return nil, err
 	}
@@ -153,7 +154,7 @@ func (d *deviceRestClient) DevicesByLabel(label string, ctx context.Context) ([]
 }
 
 func (d *deviceRestClient) DevicesForService(serviceId string, ctx context.Context) ([]models.Device, error) {
-	deviceURL, err := d.endpointer.URL(10)
+	deviceURL, err := d.client.URL(10)
 	if err != nil {
 		return nil, err
 	}
@@ -162,7 +163,7 @@ func (d *deviceRestClient) DevicesForService(serviceId string, ctx context.Conte
 }
 
 func (d *deviceRestClient) DevicesForServiceByName(serviceName string, ctx context.Context) ([]models.Device, error) {
-	deviceURL, err := d.endpointer.URL(10)
+	deviceURL, err := d.client.URL(10)
 	if err != nil {
 		return nil, err
 	}
@@ -171,7 +172,7 @@ func (d *deviceRestClient) DevicesForServiceByName(serviceName string, ctx conte
 }
 
 func (d *deviceRestClient) DevicesForProfile(profileId string, ctx context.Context) ([]models.Device, error) {
-	deviceURL, err := d.endpointer.URL(10)
+	deviceURL, err := d.client.URL(10)
 	if err != nil {
 		return nil, err
 	}
@@ -180,7 +181,7 @@ func (d *deviceRestClient) DevicesForProfile(profileId string, ctx context.Conte
 }
 
 func (d *deviceRestClient) DevicesForProfileByName(profileName string, ctx context.Context) ([]models.Device, error) {
-	deviceURL, err := d.endpointer.URL(10)
+	deviceURL, err := d.client.URL(10)
 	if err != nil {
 		return nil, err
 	}
@@ -189,7 +190,7 @@ func (d *deviceRestClient) DevicesForProfileByName(profileName string, ctx conte
 }
 
 func (d *deviceRestClient) Add(dev *models.Device, ctx context.Context) (string, error) {
-	deviceURL, err := d.endpointer.URL(10)
+	deviceURL, err := d.client.URL(10)
 	if err != nil {
 		return "", err
 	}
@@ -198,7 +199,7 @@ func (d *deviceRestClient) Add(dev *models.Device, ctx context.Context) (string,
 }
 
 func (d *deviceRestClient) Update(dev models.Device, ctx context.Context) error {
-	deviceURL, err := d.endpointer.URL(10)
+	deviceURL, err := d.client.URL(10)
 	if err != nil {
 		return err
 	}
@@ -207,7 +208,7 @@ func (d *deviceRestClient) Update(dev models.Device, ctx context.Context) error 
 }
 
 func (d *deviceRestClient) UpdateLastConnected(id string, time int64, ctx context.Context) error {
-	deviceURL, err := d.endpointer.URL(10)
+	deviceURL, err := d.client.URL(10)
 	if err != nil {
 		return err
 	}
@@ -217,7 +218,7 @@ func (d *deviceRestClient) UpdateLastConnected(id string, time int64, ctx contex
 }
 
 func (d *deviceRestClient) UpdateLastConnectedByName(name string, time int64, ctx context.Context) error {
-	deviceURL, err := d.endpointer.URL(10)
+	deviceURL, err := d.client.URL(10)
 	if err != nil {
 		return err
 	}
@@ -228,7 +229,7 @@ func (d *deviceRestClient) UpdateLastConnectedByName(name string, time int64, ct
 }
 
 func (d *deviceRestClient) UpdateLastReported(id string, time int64, ctx context.Context) error {
-	deviceURL, err := d.endpointer.URL(10)
+	deviceURL, err := d.client.URL(10)
 	if err != nil {
 		return err
 	}
@@ -238,7 +239,7 @@ func (d *deviceRestClient) UpdateLastReported(id string, time int64, ctx context
 }
 
 func (d *deviceRestClient) UpdateLastReportedByName(name string, time int64, ctx context.Context) error {
-	deviceURL, err := d.endpointer.URL(10)
+	deviceURL, err := d.client.URL(10)
 	if err != nil {
 		return err
 	}
@@ -249,7 +250,7 @@ func (d *deviceRestClient) UpdateLastReportedByName(name string, time int64, ctx
 }
 
 func (d *deviceRestClient) UpdateOpState(id string, opState string, ctx context.Context) error {
-	deviceURL, err := d.endpointer.URL(10)
+	deviceURL, err := d.client.URL(10)
 	if err != nil {
 		return err
 	}
@@ -259,7 +260,7 @@ func (d *deviceRestClient) UpdateOpState(id string, opState string, ctx context.
 }
 
 func (d *deviceRestClient) UpdateOpStateByName(name string, opState string, ctx context.Context) error {
-	deviceURL, err := d.endpointer.URL(10)
+	deviceURL, err := d.client.URL(10)
 	if err != nil {
 		return err
 	}
@@ -269,7 +270,7 @@ func (d *deviceRestClient) UpdateOpStateByName(name string, opState string, ctx 
 }
 
 func (d *deviceRestClient) UpdateAdminState(id string, adminState string, ctx context.Context) error {
-	deviceURL, err := d.endpointer.URL(10)
+	deviceURL, err := d.client.URL(10)
 	if err != nil {
 		return err
 	}
@@ -279,7 +280,7 @@ func (d *deviceRestClient) UpdateAdminState(id string, adminState string, ctx co
 }
 
 func (d *deviceRestClient) UpdateAdminStateByName(name string, adminState string, ctx context.Context) error {
-	deviceURL, err := d.endpointer.URL(10)
+	deviceURL, err := d.client.URL(10)
 	if err != nil {
 		return err
 	}
@@ -289,7 +290,7 @@ func (d *deviceRestClient) UpdateAdminStateByName(name string, adminState string
 }
 
 func (d *deviceRestClient) Delete(id string, ctx context.Context) error {
-	deviceURL, err := d.endpointer.URL(10)
+	deviceURL, err := d.client.URL(10)
 	if err != nil {
 		return err
 	}
@@ -298,7 +299,7 @@ func (d *deviceRestClient) Delete(id string, ctx context.Context) error {
 }
 
 func (d *deviceRestClient) DeleteByName(name string, ctx context.Context) error {
-	deviceURL, err := d.endpointer.URL(10)
+	deviceURL, err := d.client.URL(10)
 	if err != nil {
 		return err
 	}
